Somewhat completed the grabbing values from tables and inputting them into reports. Have not tested.
Found out solution. On Report Layout 2, The header textboxes and the first textbox on the first page are only merging. So as a solution, copy that one textbox and paste it onto where the other text boxes are and copy the contents of the the texbox into it. Then it will merge.